requirements.txt
psycopg2==2.9.5
python-dotenv==0.21.0
schedule==1.1.0
